extracted websocket code
